This branch contains the FFmpeg changes used in the Android app

+++ libavcodec/aarch64/mpegaudiodsp_init.c
@@ -16,7 +16,6 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#include <stddef.h>
 #include <stdint.h>
 
 #include "libavutil/attributes.h"
@@ -25,9 +24,9 @@
 #include "config.h"
 
 void ff_mpadsp_apply_window_fixed_neon(int32_t *synth_buf, int32_t *window,
-                                       int *dither, int16_t *samples, ptrdiff_t incr);
+                                       int *dither, int16_t *samples, int incr);
 void ff_mpadsp_apply_window_float_neon(float *synth_buf, float *window,
-                                       int *dither, float *samples, ptrdiff_t incr);
+                                       int *dither, float *samples, int incr);
 
 av_cold void ff_mpadsp_init_aarch64(MPADSPContext *s)
 {
