+++ libavcodec/eamad.c
@@ -41,8 +41,8 @@
 #include "mpeg12vlc.h"
 
 #define EA_PREAMBLE_SIZE    8
-#define MADk_TAG MKTAG('M', 'A', 'D', 'k')    /* MAD I-frame */
-#define MADm_TAG MKTAG('M', 'A', 'D', 'm')    /* MAD P-frame */
+#define MADk_TAG MKTAG('M', 'A', 'D', 'k')    /* MAD i-frame */
+#define MADm_TAG MKTAG('M', 'A', 'D', 'm')    /* MAD p-frame */
 #define MADe_TAG MKTAG('M', 'A', 'D', 'e')    /* MAD lqp-frame */
 
 typedef struct MadContext {
@@ -80,8 +80,8 @@ static av_cold int decode_init(AVCodecContext *avctx)
     return 0;
 }
 
-static inline void comp(unsigned char *dst, ptrdiff_t dst_stride,
-                        unsigned char *src, ptrdiff_t src_stride, int add)
+static inline void comp(unsigned char *dst, int dst_stride,
+                        unsigned char *src, int src_stride, int add)
 {
     int j, i;
     for (j=0; j<8; j++)
