+++ libavcodec/error_resilience.h
@@ -1,4 +1,5 @@
 /*
+ *
  * This file is part of FFmpeg.
  *
  * FFmpeg is free software; you can redistribute it and/or
@@ -41,7 +42,7 @@ typedef struct ERPicture {
     AVFrame *f;
     ThreadFrame *tf;
 
-    // it is the caller's responsibility to allocate these buffers
+    // it's the caller's responsibility to allocate these buffers
     int16_t (*motion_val[2])[2];
     int8_t *ref_index[2];
 
@@ -57,8 +58,8 @@ typedef struct ERContext {
     int *mb_index2xy;
     int mb_num;
     int mb_width, mb_height;
-    ptrdiff_t mb_stride;
-    ptrdiff_t b8_stride;
+    int mb_stride;
+    int b8_stride;
 
     volatile int error_count;
     int error_occurred;
