+++ libavcodec/iirfilter.h
@@ -27,8 +27,7 @@
 #ifndef AVCODEC_IIRFILTER_H
 #define AVCODEC_IIRFILTER_H
 
-#include <stddef.h>
-#include <stdint.h>
+#include "avcodec.h"
 
 struct FFIIRFilterCoeffs;
 struct FFIIRFilterState;
@@ -62,7 +61,7 @@ typedef struct FFIIRFilterContext {
     */
     void (*filter_flt)(const struct FFIIRFilterCoeffs *coeffs,
                         struct FFIIRFilterState *state, int size,
-                        const float *src, ptrdiff_t sstep, float *dst, ptrdiff_t dstep);
+                        const float *src, int sstep, float *dst, int dstep);
 } FFIIRFilterContext;
 
 /**
@@ -126,7 +125,7 @@ void ff_iir_filter_free_statep(struct FFIIRFilterState **state);
  * @param dstep  destination stride
  */
 void ff_iir_filter(const struct FFIIRFilterCoeffs *coeffs, struct FFIIRFilterState *state,
-                   int size, const int16_t *src, ptrdiff_t sstep, int16_t *dst, ptrdiff_t dstep);
+                   int size, const int16_t *src, int sstep, int16_t *dst, int dstep);
 
 /**
  * Perform IIR filtering on floating-point input samples.
@@ -141,7 +140,6 @@ void ff_iir_filter(const struct FFIIRFilterCoeffs *coeffs, struct FFIIRFilterSta
  */
 void ff_iir_filter_flt(const struct FFIIRFilterCoeffs *coeffs,
                        struct FFIIRFilterState *state, int size,
-                       const float *src, ptrdiff_t sstep,
-                       float *dst, ptrdiff_t dstep);
+                       const float *src, int sstep, float *dst, int dstep);
 
 #endif /* AVCODEC_IIRFILTER_H */
