+++ libavcodec/ivi_dsp.h
@@ -22,14 +22,13 @@
 
 /**
  * @file
- * DSP functions (inverse transforms, motion compensations, wavelet recomposition)
+ * DSP functions (inverse transforms, motion compensations, wavelet recompostion)
  * for Indeo Video Interactive codecs.
  */
 
 #ifndef AVCODEC_IVI_DSP_H
 #define AVCODEC_IVI_DSP_H
 
-#include <stddef.h>
 #include <stdint.h>
 
 #include "ivi.h"
@@ -42,7 +41,7 @@
  *  @param[in]   dst_pitch    pitch of the destination buffer
  */
 void ff_ivi_recompose53(const IVIPlaneDesc *plane, uint8_t *dst,
-                        const ptrdiff_t dst_pitch);
+                        const int dst_pitch);
 
 /**
  *  Haar wavelet recomposition filter for Indeo 4
@@ -52,7 +51,7 @@ void ff_ivi_recompose53(const IVIPlaneDesc *plane, uint8_t *dst,
  *  @param[in]  dst_pitch    pitch of the destination buffer
  */
 void ff_ivi_recompose_haar(const IVIPlaneDesc *plane, uint8_t *dst,
-                           const ptrdiff_t dst_pitch);
+                           const int dst_pitch);
 
 /**
  *  two-dimensional inverse Haar 8x8 transform for Indeo 4
@@ -64,7 +63,7 @@ void ff_ivi_recompose_haar(const IVIPlaneDesc *plane, uint8_t *dst,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_inverse_haar_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_inverse_haar_8x8(const int32_t *in, int16_t *out, uint32_t pitch,
                              const uint8_t *flags);
 void ff_ivi_inverse_haar_8x1(const int32_t *in, int16_t *out, uint32_t pitch,
                              const uint8_t *flags);
@@ -81,7 +80,7 @@ void ff_ivi_inverse_haar_1x8(const int32_t *in, int16_t *out, uint32_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_row_haar8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_row_haar8(const int32_t *in, int16_t *out, uint32_t pitch,
                       const uint8_t *flags);
 
 /**
@@ -94,7 +93,7 @@ void ff_ivi_row_haar8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_col_haar8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_col_haar8(const int32_t *in, int16_t *out, uint32_t pitch,
                       const uint8_t *flags);
 
 /**
@@ -107,7 +106,7 @@ void ff_ivi_col_haar8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_inverse_haar_4x4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_inverse_haar_4x4(const int32_t *in, int16_t *out, uint32_t pitch,
                              const uint8_t *flags);
 
 /**
@@ -120,7 +119,7 @@ void ff_ivi_inverse_haar_4x4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_row_haar4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_row_haar4(const int32_t *in, int16_t *out, uint32_t pitch,
                       const uint8_t *flags);
 
 /**
@@ -133,7 +132,7 @@ void ff_ivi_row_haar4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_col_haar4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_col_haar4(const int32_t *in, int16_t *out, uint32_t pitch,
                       const uint8_t *flags);
 
 /**
@@ -146,7 +145,7 @@ void ff_ivi_col_haar4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *  @param[in]  pitch       pitch to move to the next y line
  *  @param[in]  blk_size    transform block size
  */
-void ff_ivi_dc_haar_2d(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_dc_haar_2d(const int32_t *in, int16_t *out, uint32_t pitch,
                        int blk_size);
 
 /**
@@ -159,7 +158,7 @@ void ff_ivi_dc_haar_2d(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_inverse_slant_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_inverse_slant_8x8(const int32_t *in, int16_t *out, uint32_t pitch,
                               const uint8_t *flags);
 
 /**
@@ -172,7 +171,7 @@ void ff_ivi_inverse_slant_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_inverse_slant_4x4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_inverse_slant_4x4(const int32_t *in, int16_t *out, uint32_t pitch,
                               const uint8_t *flags);
 
 /**
@@ -186,7 +185,7 @@ void ff_ivi_inverse_slant_4x4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *  @param[in]    pitch       pitch to move to the next y line
  *  @param[in]    blk_size    transform block size
  */
-void ff_ivi_dc_slant_2d(const int32_t *in, int16_t *out, ptrdiff_t pitch, int blk_size);
+void ff_ivi_dc_slant_2d(const int32_t *in, int16_t *out, uint32_t pitch, int blk_size);
 
 /**
  *  inverse 1D row slant transform
@@ -196,7 +195,7 @@ void ff_ivi_dc_slant_2d(const int32_t *in, int16_t *out, ptrdiff_t pitch, int bl
  *  @param[in]    pitch   pitch to move to the next y line
  *  @param[in]    flags   pointer to the array of column flags (unused here)
  */
-void ff_ivi_row_slant8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_row_slant8(const int32_t *in, int16_t *out, uint32_t pitch,
                        const uint8_t *flags);
 
 /**
@@ -209,7 +208,7 @@ void ff_ivi_row_slant8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_col_slant8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_col_slant8(const int32_t *in, int16_t *out, uint32_t pitch,
                        const uint8_t *flags);
 
 /**
@@ -220,7 +219,7 @@ void ff_ivi_col_slant8(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *  @param[in]    pitch   pitch to move to the next y line
  *  @param[in]    flags   pointer to the array of column flags (unused here)
  */
-void ff_ivi_row_slant4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_row_slant4(const int32_t *in, int16_t *out, uint32_t pitch,
                        const uint8_t *flags);
 
 /**
@@ -233,29 +232,29 @@ void ff_ivi_row_slant4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
  *                        != 0 - non_empty column, 0 - empty one
  *                        (this array must be filled by caller)
  */
-void ff_ivi_col_slant4(const int32_t *in, int16_t *out, ptrdiff_t pitch,
+void ff_ivi_col_slant4(const int32_t *in, int16_t *out, uint32_t pitch,
                        const uint8_t *flags);
 
 /**
  *  DC-only inverse row slant transform
  */
-void ff_ivi_dc_row_slant(const int32_t *in, int16_t *out, ptrdiff_t pitch, int blk_size);
+void ff_ivi_dc_row_slant(const int32_t *in, int16_t *out, uint32_t pitch, int blk_size);
 
 /**
  *  DC-only inverse column slant transform
  */
-void ff_ivi_dc_col_slant(const int32_t *in, int16_t *out, ptrdiff_t pitch, int blk_size);
+void ff_ivi_dc_col_slant(const int32_t *in, int16_t *out, uint32_t pitch, int blk_size);
 
 /**
  *  Copy the pixels into the frame buffer.
  */
-void ff_ivi_put_pixels_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch, const uint8_t *flags);
+void ff_ivi_put_pixels_8x8(const int32_t *in, int16_t *out, uint32_t pitch, const uint8_t *flags);
 
 /**
  *  Copy the DC coefficient into the first pixel of the block and
  *  zero all others.
  */
-void ff_ivi_put_dc_pixel_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch, int blk_size);
+void ff_ivi_put_dc_pixel_8x8(const int32_t *in, int16_t *out, uint32_t pitch, int blk_size);
 
 /**
  *  8x8 block motion compensation with adding delta
@@ -265,7 +264,7 @@ void ff_ivi_put_dc_pixel_8x8(const int32_t *in, int16_t *out, ptrdiff_t pitch, i
  *  @param[in]       pitch    pitch for moving to the next y line
  *  @param[in]       mc_type  interpolation type
  */
-void ff_ivi_mc_8x8_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, int mc_type);
+void ff_ivi_mc_8x8_delta(int16_t *buf, const int16_t *ref_buf, uint32_t pitch, int mc_type);
 
 /**
  *  4x4 block motion compensation with adding delta
@@ -275,7 +274,7 @@ void ff_ivi_mc_8x8_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, 
  *  @param[in]       pitch    pitch for moving to the next y line
  *  @param[in]       mc_type  interpolation type
  */
-void ff_ivi_mc_4x4_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, int mc_type);
+void ff_ivi_mc_4x4_delta(int16_t *buf, const int16_t *ref_buf, uint32_t pitch, int mc_type);
 
 /**
  *  motion compensation without adding delta
@@ -285,7 +284,7 @@ void ff_ivi_mc_4x4_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, 
  *  @param[in]      pitch    pitch for moving to the next y line
  *  @param[in]      mc_type  interpolation type
  */
-void ff_ivi_mc_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, int mc_type);
+void ff_ivi_mc_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, uint32_t pitch, int mc_type);
 
 /**
  *  4x4 block motion compensation without adding delta
@@ -295,7 +294,7 @@ void ff_ivi_mc_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitc
  *  @param[in]      pitch    pitch for moving to the next y line
  *  @param[in]      mc_type  interpolation type
  */
-void ff_ivi_mc_4x4_no_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitch, int mc_type);
+void ff_ivi_mc_4x4_no_delta(int16_t *buf, const int16_t *ref_buf, uint32_t pitch, int mc_type);
 
 /**
  *  8x8 block motion compensation with adding delta
@@ -307,7 +306,7 @@ void ff_ivi_mc_4x4_no_delta(int16_t *buf, const int16_t *ref_buf, ptrdiff_t pitc
  *  @param[in]      mc_type  interpolation type for backward reference
  *  @param[in]      mc_type2 interpolation type for forward reference
  */
-void ff_ivi_mc_avg_8x8_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, ptrdiff_t pitch, int mc_type, int mc_type2);
+void ff_ivi_mc_avg_8x8_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, uint32_t pitch, int mc_type, int mc_type2);
 
 /**
  *  4x4 block motion compensation with adding delta
@@ -319,7 +318,7 @@ void ff_ivi_mc_avg_8x8_delta(int16_t *buf, const int16_t *ref_buf, const int16_t
  *  @param[in]      mc_type  interpolation type for backward reference
  *  @param[in]      mc_type2 interpolation type for forward reference
  */
-void ff_ivi_mc_avg_4x4_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, ptrdiff_t pitch, int mc_type, int mc_type2);
+void ff_ivi_mc_avg_4x4_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, uint32_t pitch, int mc_type, int mc_type2);
 
 /**
  *  motion compensation without adding delta for B-frames
@@ -331,7 +330,7 @@ void ff_ivi_mc_avg_4x4_delta(int16_t *buf, const int16_t *ref_buf, const int16_t
  *  @param[in]      mc_type  interpolation type for backward reference
  *  @param[in]      mc_type2 interpolation type for forward reference
  */
-void ff_ivi_mc_avg_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, ptrdiff_t pitch, int mc_type, int mc_type2);
+void ff_ivi_mc_avg_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, uint32_t pitch, int mc_type, int mc_type2);
 
 /**
  *  4x4 block motion compensation without adding delta for B-frames
@@ -343,6 +342,6 @@ void ff_ivi_mc_avg_8x8_no_delta(int16_t *buf, const int16_t *ref_buf, const int1
  *  @param[in]      mc_type  interpolation type for backward reference
  *  @param[in]      mc_type2 interpolation type for forward reference
  */
-void ff_ivi_mc_avg_4x4_no_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, ptrdiff_t pitch, int mc_type, int mc_type2);
+void ff_ivi_mc_avg_4x4_no_delta(int16_t *buf, const int16_t *ref_buf, const int16_t *ref_buf2, uint32_t pitch, int mc_type, int mc_type2);
 
 #endif /* AVCODEC_IVI_DSP_H */
